# Conflicts:
#	.github/ci/files/var/classes/DataObject/FilterDefinition.php
#	.github/ci/files/var/classes/DataObject/OnlineShopOrder.php
#	.github/ci/files/var/classes/DataObject/OnlineShopOrderItem.php
#	.github/ci/files/var/classes/DataObject/OnlineShopTaxClass.php
#	.github/ci/files/var/classes/DataObject/OnlineShopVoucherSeries.php
#	.github/ci/files/var/classes/DataObject/OnlineShopVoucherToken.php
#	bundles/AdminBundle/public/js/pimcore/object/folder.js
#	bundles/AdminBundle/public/js/pimcore/object/object.js
#	bundles/AdminBundle/src/Controller/Admin/DataObject/DataObjectController.php
#	bundles/AdminBundle/src/Controller/Admin/TagsController.php
#	bundles/AdminBundle/src/Helper/QueryParams.php
#	bundles/EcommerceFrameworkBundle/src/Controller/AdminOrderController.php
#	bundles/EcommerceFrameworkBundle/src/IndexService/Config/ElasticSearch.php
#	bundles/EcommerceFrameworkBundle/src/IndexService/Worker/DefaultFindologic.php
#	bundles/EcommerceFrameworkBundle/src/IndexService/Worker/OptimizedMysql.php
#	doc/Development_Documentation/05_Objects/01_Object_Classes/05_Class_Settings/55_Preview.md
#	doc/Development_Documentation/23_Installation_and_Upgrade/09_Upgrade_Notes/README.md
#	models/DataObject/AbstractObject.php
#	models/DataObject/Concrete.php
#	models/DataObject/Folder.php
#	models/DataObject/Objectbrick/Data/Dao.php
#	models/DataObject/Service.php
#	models/Element/Traits/DirtyIndicatorTrait.php
#	tests/Support/Helper/AbstractTestDataHelper.php

+++ bundles/CoreBundle/src/Migrations/Version20201008101817.php
@@ -1,4 +1,5 @@
 <?php
+declare(strict_types=1);
 
 /**
  * Pimcore
@@ -23,21 +24,17 @@ use Doctrine\Migrations\AbstractMigration;
  */
 final class Version20201008101817 extends AbstractMigration
 {
-    /**
-     * @param Schema $schema
-     */
     public function up(Schema $schema): void
     {
-        $this->addSql('DROP VIEW IF EXISTS documents_editables;');
+        if (array_key_exists('documents_editables', $this->connection->getSchemaManager()->listViews())) {
+            $this->addSql('DROP VIEW IF EXISTS documents_editables;');
+        }
 
         if ($schema->hasTable('documents_elements')) {
             $this->addSql('RENAME TABLE documents_elements TO documents_editables;');
         }
     }
 
-    /**
-     * @param Schema $schema
-     */
     public function down(Schema $schema): void
     {
         $this->addSql('RENAME TABLE documents_editables TO documents_elements;');
